More validations

+++ src/main/java/io/github/enzopavani/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.enzopavani.libraryapi.service;
 
+import io.github.enzopavani.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.enzopavani.libraryapi.model.Autor;
 import io.github.enzopavani.libraryapi.repository.AutorRepository;
+import io.github.enzopavani.libraryapi.repository.LivroRepository;
 import io.github.enzopavani.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +16,12 @@ public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -30,6 +34,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir Autor que possui livros cadastrados");
+        }
         repository.delete(autor);
     }
 
@@ -52,5 +60,9 @@ public class AutorService {
         }
         validator.validar(autor);
         repository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
